+++ src/App/FileInputs.fs
@@ -14,16 +14,16 @@ let view() =
     let files = Store.make Unchecked.defaultof<Browser.Types.FileList>
     let fileSeq = files |> Store.map (Helpers.fileListToSeq >> Seq.toList)
 
-    fileSeq |> (Store.iter (fun fileSeq ->
+    let unsub = fileSeq |> (Store.iter (fun fileSeq ->
         // Note that `files` is of type `FileList`, not an Array:
         // https://developer.mozilla.org/en-US/docs/Web/API/FileList
         //console.log(files);
 
         for file in fileSeq do
-            Browser.Dom.console.log($"{file.name}: {file.size} bytes")))
+            Browser.Dom.console.log($"{file.name}: {file.size} bytes"))) 
 
     Html.div [
-        disposeOnUnmount [ files ]
+        disposeOnUnmount [ files; unsub ]
 
         Html.div [
             class' "block"