+++ src/App/ReactiveStatements.fs
@@ -15,7 +15,7 @@ let plural n = if n = 1 then "" else "s"
 let view() =
     let count = Store.make 0
 
-    count |> Store.iter (fun n ->
+    let unsub = count |> Store.iter (fun n ->
         if n >= 10 then
             window.alert("count is dangerously high!")
             count <~ 9
@@ -25,7 +25,7 @@ let view() =
         count <~= inc   // or: Store.modify count inc
 
     Html.button [
-        disposeOnUnmount [count]
+        disposeOnUnmount [count; unsub ]
 
         onClick handleClick []