+++ src/Sutil/ConstructStyleSheetsPolyfill.fs
@@ -1,0 +1,7 @@
+module ConstructStyleSheetsPolyfill
+
+open Fable.Core.JsInterop
+
+importSideEffects "./adoptedStyleSheets.js"
+
+let register() = ()